+++ CHANGES.rst
@@ -11,6 +11,10 @@ https://github.com/zopefoundation/Zope/blob/4.x/CHANGES.rst
 5.8 (unreleased)
 ----------------
 
+- Only set response header Content-Type as text/html on exception views when
+  the response has content.
+  (`#1089 <https://github.com/zopefoundation/Zope/issues/1089>`_)
+
 - Drop support for Python 3.6, it has been in end-of-life status for a while.
 
 - Update to newest compatible versions of dependencies.